Disable invalid O_NOFOLLOW on PASE
AIX messed up O_NOFOLLOW by setting it to a flag that's greater
than can be expressed as an integer. This was corrected in later
versions of AIX, but PASE 7.2 and 7.3 still have the incorrect
values. If O_NOFOLLOW is on the os package, the tempfile module
will automatically try to use it on the flags it passes to open,
which will cause an OverflowError to occur.

Any user of the tempfile package is broken.